// src/ui/Typography/Typography.types.ts
export const enum ETypographyType {
  title_27_400_primary_037 = 'title_27_400_primary_037',
  title_24_600_primary = 'title_24_600_primary',
  title_20_600_primary = 'title_20_600_primary',
  title_20_600_primary_087 = 'title_20_600_primary_087',
  subtitle_16_700_primary_087 = 'subtitle_16_700_primary_087',
  subtitle_16_600_primary = 'subtitle_16_600_primary',
  subtitle_16_600_primary_087 = 'subtitle_16_600_primary_087',
  subtitle_16_400_primary = 'subtitle_16_400_primary',
  subtitle_16_500_primary_087 = 'subtitle_16_500_primary_087',
  subtitle_16_400_primary_087 = 'subtitle_16_400_primary_087',
  subtitle_16_400_primary_037 = 'subtitle_16_400_primary_037',
  subtitle_16_400_grey_500 = 'subtitle_16_400_grey_500',
  text_14_400_primary_087 = 'text_14_400_primary_087',
  text_14_500_primary_06 = 'text_14_500_primary_06',
  text_14_400_primary_06 = 'text_14_400_primary_06',
  text_12_400_primary_087 = 'text_12_400_primary_087',
  text_12_400_primary_06 = 'text_12_400_primary_06',
  text_10_400_secondary = 'text_10_400_secondary',
  text_10_400_black = 'text_10_400_black',
  text_10_400_white = 'text_10_400_white',
  text_6_400_secondary = 'text_6_400_secondary',
}